renamed .env and updated test.yml
I have tried to run Github actions locally usuing `act`, which requires me to rename or remove .env as it is a default file name to configure the environment, so I have renamed it to config.js.

In addition, I have updated test.yml, which contains the GitHub action setup for CI. I am still testing it to see if GitHub actions work or not.
