--- .github/workflows/testing.yml
@@ -1,44 +1,0 @@
-name: Tests CI
-on: [push, pull_request]
-jobs:
-    test:
-        runs-on: ubuntu-latest
-        steps:
-            - uses: actions/checkout@v2
-            - name: Test using Node.js
-              uses: actions/setup-node@v1
-              with:
-                  node-version: "12"
-            - run: yarn install
-            - name: Create config.json
-            - uses: DamianReeves/write-file-action
-              with:
-                  path: ${{ env.home}}/config.json
-                  contents: |
-                      `{
-                          "BITLY_TOKEN": "${{ secrets.BITLY_TOKEN }}"
-                      }`
-                  write-mode: append
-            - run: yarn test:ci
-            - name: Tests ✅
-              if: ${{ success() }}
-              run: |
-                  curl --request POST           --url https://api.github.com/repos/${{ github.repository }}/statuses/${{ github.sha }}           --header 'authorization: Bearer ${{ secrets.GITHUB_TOKEN }}'           --header 'content-type: application/json'           --data '{
-                    "context": "tests",
-                    "state": "success",
-                    "description": "Tests passed",
-                    "target_url": "https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }}"
-                  }'
-            - name: Tests 🚨
-              if: ${{ failure() }}
-              run: |
-                  curl --request POST           --url https://api.github.com/repos/${{ github.repository }}/statuses/${{ github.sha }}           --header 'authorization: Bearer ${{ secrets.GITHUB_TOKEN }}'           --header 'content-type: application/json'           --data '{
-                    "context": "tests",
-                    "state": "failure",
-                    "description": "Tests failed",
-                    "target_url": "https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }}"
-                  }'
-            - name: Remove File
-              uses: JesseTG/rm@v1.0.2
-              with:
-                  path: ${{ env.home}}/config.json
