+++ docs/js/docma-web.js
@@ -6589,7 +6589,7 @@ dust.filters.$id = function (symbol) {
 DocmaWeb.version = "3.2.2";
 return DocmaWeb;
 })();
-var docma = Object.freeze(new DocmaWeb({"version":"3.2.2","routes":[{"id":"api:","type":"api","name":"_def_","path":"?api","contentPath":null,"sourceType":"js"},{"id":"content:code_of_conduct","type":"content","name":"code_of_conduct","path":"?content=code_of_conduct","contentPath":"content/code_of_conduct.html","sourceType":"md"},{"id":"content:readme","type":"content","name":"readme","path":"?content=readme","contentPath":"content/readme.html","sourceType":"md"},{"id":"content:security","type":"content","name":"security","path":"?content=security","contentPath":"content/security.html","sourceType":"md"}],"apis":{"_def_":{"documentation":[{"comment":"/**\n * Execute on addembedding command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[138,300],"filename":"addembedding.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100000953","name":"addembedding","type":"ArrowFunctionExpression"},"vars":{"embed":"addembedding~embed"}},"description":"Execute on addembedding command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"addembedding","longname":"addembedding","kind":"function","scope":"global","$longname":"addembedding","$kind":"method","$docmaLink":"?api#addembedding"},{"comment":"/**\n * Execute on addproject command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[139,1142],"filename":"addproject.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100000994","name":"addproject","type":"ArrowFunctionExpression"},"vars":{"roles":"addproject~roles","channels":"addproject~channels","":null,"project":"addproject~project"}},"description":"Execute on addproject command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"addproject","longname":"addproject","kind":"function","scope":"global","async":true,"$longname":"addproject","$kind":"method","$docmaLink":"?api#addproject"},{"comment":"/**\n * Execute on addrole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[136,1173],"filename":"addrole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001148","name":"addrole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"addrole~givenRole","member":"addrole~member","findRole":"addrole~findRole","":null}},"description":"Execute on addrole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"addrole","longname":"addrole","kind":"function","scope":"global","async":true,"$longname":"addrole","$kind":"method","$docmaLink":"?api#addrole"},{"comment":"/**\n * Execute on adds command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[181,1032],"filename":"adds.js","lineno":7,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001311","name":"adds","type":"ArrowFunctionExpression"},"vars":{"twitch_username":"adds~twitch_username","discord_username":"adds~discord_username","discord_id":"adds~discord_id","twitch_api":"adds~twitch_api","twitch_id":"adds~twitch_id","discord_guild":"adds~discord_guild"}},"description":"Execute on adds command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"adds","longname":"adds","kind":"function","scope":"global","async":true,"$longname":"adds","$kind":"method","$docmaLink":"?api#adds"},{"comment":"/**\n * Execute on addstreamer command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[140,1535],"filename":"addstreamer.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001426","name":"addstreamer","type":"ArrowFunctionExpression"},"vars":{"roles":"addstreamer~roles","channels":"addstreamer~channels","":null,"twitch_username":"addstreamer~twitch_username","discord_username":"addstreamer~discord_username"}},"description":"Execute on addstreamer command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"addstreamer","longname":"addstreamer","kind":"function","scope":"global","async":true,"$longname":"addstreamer","$kind":"method","$docmaLink":"?api#addstreamer"},{"comment":"/**\n * Execute on args command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[133,318],"filename":"args.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001627","name":"args","type":"ArrowFunctionExpression"}},"description":"Execute on args command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"args","longname":"args","kind":"function","scope":"global","async":true,"$longname":"args","$kind":"method","$docmaLink":"?api#args"},{"comment":"/**\n * Execute on avatar command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[93,704],"filename":"avatar.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001678","name":"avatar","type":"ArrowFunctionExpression"},"vars":{"avatarList":"avatar~avatarList","":null}},"description":"Execute on avatar command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"avatar","longname":"avatar","kind":"function","scope":"global","async":true,"$longname":"avatar","$kind":"method","$docmaLink":"?api#avatar"},{"comment":"/**\n * Execute on ban command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[90,232],"filename":"ban.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001774","name":"ban","type":"ArrowFunctionExpression"},"vars":{"user":"ban~user"}},"description":"Execute on ban command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"ban","longname":"ban","kind":"function","scope":"global","async":true,"$longname":"ban","$kind":"method","$docmaLink":"?api#ban"},{"comment":"/**\n * Execute on clear command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[92,824],"filename":"clear.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001829","name":"clear","type":"ArrowFunctionExpression"},"vars":{"deleteMessages":"clear~deleteMessages"}},"description":"Execute on clear command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"clear","longname":"clear","kind":"function","scope":"global","async":true,"$longname":"clear","$kind":"method","$docmaLink":"?api#clear"},{"comment":"/**\n * Execute on getdatabase command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[98,425],"filename":"getdatabase.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001938","name":"getdatabase","type":"ArrowFunctionExpression"},"vars":{"streamerDB":"getdatabase~streamerDB"}},"description":"Execute on getdatabase command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"getdatabase","longname":"getdatabase","kind":"function","scope":"global","async":true,"$longname":"getdatabase","$kind":"method","$docmaLink":"?api#getdatabase"},{"comment":"/**\n * Execute on getstreamers command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[141,1514],"filename":"getstreamers.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001976","name":"getstreamers","type":"ArrowFunctionExpression"},"vars":{"roles":"getstreamers~roles","channels":"getstreamers~channels","":null,"twitch_username":"getstreamers~twitch_username","discord_username":"getstreamers~discord_username"}},"description":"Execute on getstreamers command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"getstreamers","longname":"getstreamers","kind":"function","scope":"global","async":true,"$longname":"getstreamers","$kind":"method","$docmaLink":"?api#getstreamers"},{"comment":"/**\n * Get User\n * @param {String} username - twitch username\n */","meta":{"range":[790,1199],"filename":"Twitch.js","lineno":31,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/utils","code":{"id":"astnode100006517","name":"getUser","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Get User","params":[{"type":{"names":["String"]},"description":"twitch username","name":"username"}],"name":"getUser","longname":"getUser","kind":"function","scope":"global","async":true,"$longname":"getUser","$kind":"method","$docmaLink":"?api#getUser"},{"comment":"/**\n * Execute on help command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[175,1791],"filename":"help.js","lineno":7,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002175","name":"help","type":"ArrowFunctionExpression"},"vars":{"data":"help~data","undefined":null,"":null,"name":"help~name","command":"help~command"}},"description":"Execute on help command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"help","longname":"help","kind":"function","scope":"global","async":true,"$longname":"help","$kind":"method","$docmaLink":"?api#help"},{"comment":"/**\n * Execute on invite command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[135,954],"filename":"invite.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002461","name":"invite","type":"ArrowFunctionExpression"},"vars":{"time":"invite~time","usage":"invite~usage","invite":"invite~invite"}},"description":"Execute on invite command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"invite","longname":"invite","kind":"function","scope":"global","async":true,"$longname":"invite","$kind":"method","$docmaLink":"?api#invite"},{"comment":"/**\n * Is Live\n * @param {Twitch.User} user - twitch user\n */","meta":{"range":[152,716],"filename":"Twitch.js","lineno":8,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/utils","code":{"id":"astnode100006445","name":"isLive","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Is Live","params":[{"type":{"names":["Twitch.User"]},"description":"twitch user","name":"user"}],"name":"isLive","longname":"isLive","kind":"function","scope":"global","async":true,"$longname":"isLive","$kind":"method","$docmaLink":"?api#isLive"},{"comment":"/**\n * Execute on kick command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[91,235],"filename":"kick.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002566","name":"kick","type":"ArrowFunctionExpression"},"vars":{"user":"kick~user"}},"description":"Execute on kick command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"kick","longname":"kick","kind":"function","scope":"global","async":true,"$longname":"kick","$kind":"method","$docmaLink":"?api#kick"},{"comment":"/**\n * Execute on message events\n * @param {Discord.Client} client - discord client\n */","meta":{"range":[216,5280],"filename":"message.js","lineno":9,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/actions","code":{"id":"astnode100000022","name":"message","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on message events","params":[{"type":{"names":["Discord.Client"]},"description":"discord client","name":"client"}],"name":"message","longname":"message","kind":"function","scope":"global","$longname":"message","$kind":"method","$docmaLink":"?api#message"},{"comment":"/**\n * Execute on mkcha command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,2630],"filename":"mkcha.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002621","name":"mkcha","type":"ArrowFunctionExpression"},"vars":{"channels":"mkcha~channels","categories":"mkcha~categories","":null,"channel_name":"mkcha~channel_name","channel_type":"mkcha~channel_type","parent_category":"mkcha~parent_category","shouldIBreak":"mkcha~shouldIBreak"}},"description":"Execute on mkcha command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"mkcha","longname":"mkcha","kind":"function","scope":"global","async":true,"$longname":"mkcha","$kind":"method","$docmaLink":"?api#mkcha"},{"comment":"/**\n * Execute on mute command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[91,570],"filename":"mute.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002886","name":"mute","type":"ArrowFunctionExpression"},"vars":{"roles":"mute~roles","":null,"user":"mute~user","muteRole":"mute~muteRole"}},"description":"Execute on mute command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"mute","longname":"mute","kind":"function","scope":"global","async":true,"$longname":"mute","$kind":"method","$docmaLink":"?api#mute"},{"comment":"/**\n * Execute on mvcha command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,2208],"filename":"mvcha.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002994","name":"mvcha","type":"ArrowFunctionExpression"},"vars":{"channel_name":"mvcha~channel_name","new_channel_name":"mvcha~new_channel_name","new_category":"mvcha~new_category","channel_id":"mvcha~channel_id","":null,"givenCategory":"mvcha~givenCategory"}},"description":"Execute on mvcha command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"mvcha","longname":"mvcha","kind":"function","scope":"global","async":true,"$longname":"mvcha","$kind":"method","$docmaLink":"?api#mvcha"},{"comment":"/**\n * Execute on newrole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[136,4510],"filename":"newrole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003201","name":"newrole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"newrole~givenRole","permissions":"newrole~permissions","findRole":"newrole~findRole","":null,"options":"newrole~options"}},"description":"Execute on newrole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"newrole","longname":"newrole","kind":"function","scope":"global","async":true,"$longname":"newrole","$kind":"method","$docmaLink":"?api#newrole"},{"comment":"/**\n * Execute on pause command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[147,735],"filename":"pause.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003454","name":"pause","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"pause~serverQueue","serverQueue.playing":"pause~serverQueue.playing"}},"description":"Execute on pause command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"pause","longname":"pause","kind":"function","scope":"global","async":true,"$longname":"pause","$kind":"method","$docmaLink":"?api#pause"},{"comment":"/**\n * Execute on pin command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[132,563],"filename":"pin.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003575","name":"pin","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on pin command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"pin","longname":"pin","kind":"function","scope":"global","async":true,"$longname":"pin","$kind":"method","$docmaLink":"?api#pin"},{"comment":"/**\n * Execute on ping command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[91,153],"filename":"ping.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003671","name":"ping","type":"ArrowFunctionExpression"}},"description":"Execute on ping command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"ping","longname":"ping","kind":"function","scope":"global","async":true,"$longname":"ping","$kind":"method","$docmaLink":"?api#ping"},{"comment":"/**\n * Execute on play command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[1354,3616],"filename":"play.js","lineno":47,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003869","name":"play","type":"ArrowFunctionExpression"},"vars":{"url":"play~url","serverQueue":"play~serverQueue","voiceChannel":"play~voiceChannel","permissions":"play~permissions","songInfo":"play~songInfo","song":"play~song","queueConstract":"play~queueConstract","connection":"play~connection","queueConstract.connection":"play~queueConstract.connection"}},"description":"Execute on play command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"play","longname":"play","kind":"function","scope":"global","async":true,"$longname":"play","$kind":"method","$docmaLink":"?api#play"},{"comment":"/**\n * Play action\n * @param {Discord.Guild} guild - discord guild\n * @param {Queue} song - song queue\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[252,1218],"filename":"play.js","lineno":10,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003715","name":"playAction","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"playAction~serverQueue","dispatcher":"playAction~dispatcher","":null}},"description":"Play action","params":[{"type":{"names":["Discord.Guild"]},"description":"discord guild","name":"guild"},{"type":{"names":["Queue"]},"description":"song queue","name":"song"},{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"playAction","longname":"playAction","kind":"function","scope":"global","$longname":"playAction","$kind":"method","$docmaLink":"?api#playAction"},{"comment":"/**\n * Execute on prune command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,636],"filename":"prune.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004176","name":"prune","type":"ArrowFunctionExpression"},"vars":{"amount":"prune~amount","":null}},"description":"Execute on prune command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"prune","longname":"prune","kind":"function","scope":"global","async":true,"$longname":"prune","$kind":"method","$docmaLink":"?api#prune"},{"comment":"/**\n * Execute on reaction events\n * @param {Discord.Client} client - discord client\n */","meta":{"range":[95,3004],"filename":"reactions.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/actions","code":{"id":"astnode100000590","name":"reactions","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on reaction events","params":[{"type":{"names":["Discord.Client"]},"description":"discord client","name":"client"}],"name":"reactions","longname":"reactions","kind":"function","scope":"global","$longname":"reactions","$kind":"method","$docmaLink":"?api#reactions"},{"comment":"/**\n * Execute on ready event\n * @param {Discord.Client} client - discord client\n */","meta":{"range":[176,2030],"filename":"ready.js","lineno":7,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/actions","code":{"id":"astnode100000887","name":"ready","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on ready event","params":[{"type":{"names":["Discord.Client"]},"description":"discord client","name":"client"}],"name":"ready","longname":"ready","kind":"function","scope":"global","$longname":"ready","$kind":"method","$docmaLink":"?api#ready"},{"comment":"/**\n * Execute on repeat command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[148,756],"filename":"repeat.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004270","name":"repeat","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"repeat~serverQueue","serverQueue.loop":"repeat~serverQueue.loop"}},"description":"Execute on repeat command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"repeat","longname":"repeat","kind":"function","scope":"global","async":true,"$longname":"repeat","$kind":"method","$docmaLink":"?api#repeat"},{"comment":"/**\n * Execute on resume command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[148,739],"filename":"resume.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004377","name":"resume","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"resume~serverQueue","serverQueue.playing":"resume~serverQueue.playing"}},"description":"Execute on resume command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"resume","longname":"resume","kind":"function","scope":"global","async":true,"$longname":"resume","$kind":"method","$docmaLink":"?api#resume"},{"comment":"/**\n * Execute on rmcha command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,699],"filename":"rmcha.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004497","name":"rmcha","type":"ArrowFunctionExpression"},"vars":{"channel_name":"rmcha~channel_name","channel_id":"rmcha~channel_id","":null,"fetchedChannel":"rmcha~fetchedChannel"}},"description":"Execute on rmcha command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"rmcha","longname":"rmcha","kind":"function","scope":"global","async":true,"$longname":"rmcha","$kind":"method","$docmaLink":"?api#rmcha"},{"comment":"/**\n * Execute on rmrole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[135,1179],"filename":"rmrole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004599","name":"rmrole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"rmrole~givenRole","member":"rmrole~member","findRole":"rmrole~findRole","":null}},"description":"Execute on rmrole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"rmrole","longname":"rmrole","kind":"function","scope":"global","async":true,"$longname":"rmrole","$kind":"method","$docmaLink":"?api#rmrole"},{"comment":"/**\n * Execute on server command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[93,243],"filename":"server.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004754","name":"server","type":"ArrowFunctionExpression"}},"description":"Execute on server command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"server","longname":"server","kind":"function","scope":"global","async":true,"$longname":"server","$kind":"method","$docmaLink":"?api#server"},{"comment":"/**\n * Execute on short command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[326,596],"filename":"short.js","lineno":10,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004825","name":"short","type":"ArrowFunctionExpression"},"vars":{"url":"short~url","response":"short~response"}},"description":"Execute on short command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"short","longname":"short","kind":"function","scope":"global","async":true,"$longname":"short","$kind":"method","$docmaLink":"?api#short"},{"comment":"/**\n * Execute on skip command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[146,687],"filename":"skip.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004894","name":"skip","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"skip~serverQueue","serverQueue.loop":"skip~serverQueue.loop"}},"description":"Execute on skip command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"skip","longname":"skip","kind":"function","scope":"global","async":true,"$longname":"skip","$kind":"method","$docmaLink":"?api#skip"},{"comment":"/**\n * Execute on song command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[146,511],"filename":"song.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005021","name":"song","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"song~serverQueue"}},"description":"Execute on song command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"song","longname":"song","kind":"function","scope":"global","async":true,"$longname":"song","$kind":"method","$docmaLink":"?api#song"},{"comment":"/**\n * Execute on stop command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[146,721],"filename":"stop.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005113","name":"stop","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"stop~serverQueue","serverQueue.songs":"stop~serverQueue.songs"}},"description":"Execute on stop command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"stop","longname":"stop","kind":"function","scope":"global","async":true,"$longname":"stop","$kind":"method","$docmaLink":"?api#stop"},{"comment":"/**\n * Execute on unban command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,716],"filename":"unban.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005246","name":"unban","type":"ArrowFunctionExpression"},"vars":{"username":"unban~username","":null}},"description":"Execute on unban command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"unban","longname":"unban","kind":"function","scope":"global","async":true,"$longname":"unban","$kind":"method","$docmaLink":"?api#unban"},{"comment":"/**\n * Execute on unmute command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[93,581],"filename":"unmute.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005348","name":"unmute","type":"ArrowFunctionExpression"},"vars":{"roles":"unmute~roles","":null,"user":"unmute~user","muteRole":"unmute~muteRole"}},"description":"Execute on unmute command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"unmute","longname":"unmute","kind":"function","scope":"global","async":true,"$longname":"unmute","$kind":"method","$docmaLink":"?api#unmute"},{"comment":"/**\n * Execute on unpin command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,1976],"filename":"unpin.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005454","name":"unpin","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on unpin command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"unpin","longname":"unpin","kind":"function","scope":"global","async":true,"$longname":"unpin","$kind":"method","$docmaLink":"?api#unpin"},{"comment":"/**\n * Execute on virole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[135,9048],"filename":"virole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005672","name":"virole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"virole~givenRole","option":"virole~option","update":"virole~update","findRole":"virole~findRole","":null,"options":"virole~options","colorHelp":"virole~colorHelp","updateHelp":"virole~updateHelp","nameHelp":"virole~nameHelp"}},"description":"Execute on virole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"virole","longname":"virole","kind":"function","scope":"global","async":true,"$longname":"virole","$kind":"method","$docmaLink":"?api#virole"}],"symbols":["addembedding","addproject","addrole","adds","addstreamer","args","avatar","ban","clear","getdatabase","getstreamers","getUser","help","invite","isLive","kick","message","mkcha","mute","mvcha","newrole","pause","pin","ping","play","playAction","prune","reactions","ready","repeat","resume","rmcha","rmrole","server","short","skip","song","stop","unban","unmute","unpin","virole"]}},"app":{"base":"/SudoBot/","title":"","meta":null,"entrance":"api","routing":{"method":"query","caseSensitive":true},"server":"static","favicon":""},"template":{"name":"docma-template-zebra","description":"Zebra - Default template for Docma. https://github.com/onury/docma","version":"2.3.1","supportedDocmaVersion":">=2.0.0","author":"Onur Yıldırım","license":"MIT","mainHTML":"index.html","options":{"title":{"label":"SudoBot","href":"/SudoBot/"},"logo":{"dark":"https://avatars.githubusercontent.com/u/72315676?s=200&v=4","light":"https://avatars.githubusercontent.com/u/72315676?s=200&v=4"},"sidebar":{"enabled":true,"outline":"tree","collapsed":false,"toolbar":true,"itemsFolded":false,"itemsOverflow":"crop","badges":true,"search":true,"animations":true},"symbols":{"autoLink":true,"params":"list","enums":"list","props":"list","meta":true},"contentView":{"bookmarks":"h1,h2,h3","faLibs":"solid,regular,brands","faVersion":"5.5.0"},"navbar":{"enabled":true,"dark":false,"animations":true,"menu":[{"label":"GitHub","iconClass":"fab fa-github","href":"https://github.com/sudogroup/SudoBot","target":"_blank"}],"fixed":true}}},"partials":{"api":"docma-api","content":"docma-content","notFound":"docma-404"},"elementID":"docma-main","contentElementID":"docma-content","defaultApiName":"_def_","logsEnabled":true}));
+var docma = Object.freeze(new DocmaWeb({"version":"3.2.2","routes":[{"id":"api:","type":"api","name":"_def_","path":"?api","contentPath":null,"sourceType":"js"},{"id":"content:code_of_conduct","type":"content","name":"code_of_conduct","path":"?content=code_of_conduct","contentPath":"content/code_of_conduct.html","sourceType":"md"},{"id":"content:readme","type":"content","name":"readme","path":"?content=readme","contentPath":"content/readme.html","sourceType":"md"},{"id":"content:security","type":"content","name":"security","path":"?content=security","contentPath":"content/security.html","sourceType":"md"}],"apis":{"_def_":{"documentation":[{"comment":"/**\n * Execute on addembedding command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[138,300],"filename":"addembedding.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100000953","name":"addembedding","type":"ArrowFunctionExpression"},"vars":{"embed":"addembedding~embed"}},"description":"Execute on addembedding command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"addembedding","longname":"addembedding","kind":"function","scope":"global","$longname":"addembedding","$kind":"method","$docmaLink":"?api#addembedding"},{"comment":"/**\n * Execute on addproject command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[139,1142],"filename":"addproject.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100000994","name":"addproject","type":"ArrowFunctionExpression"},"vars":{"roles":"addproject~roles","channels":"addproject~channels","":null,"project":"addproject~project"}},"description":"Execute on addproject command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"addproject","longname":"addproject","kind":"function","scope":"global","async":true,"$longname":"addproject","$kind":"method","$docmaLink":"?api#addproject"},{"comment":"/**\n * Execute on addrole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[136,1173],"filename":"addrole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001148","name":"addrole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"addrole~givenRole","member":"addrole~member","findRole":"addrole~findRole","":null}},"description":"Execute on addrole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"addrole","longname":"addrole","kind":"function","scope":"global","async":true,"$longname":"addrole","$kind":"method","$docmaLink":"?api#addrole"},{"comment":"/**\n * Execute on adds command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[181,1032],"filename":"adds.js","lineno":7,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001311","name":"adds","type":"ArrowFunctionExpression"},"vars":{"twitch_username":"adds~twitch_username","discord_username":"adds~discord_username","discord_id":"adds~discord_id","twitch_api":"adds~twitch_api","twitch_id":"adds~twitch_id","discord_guild":"adds~discord_guild"}},"description":"Execute on adds command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"adds","longname":"adds","kind":"function","scope":"global","async":true,"$longname":"adds","$kind":"method","$docmaLink":"?api#adds"},{"comment":"/**\n * Execute on addstreamer command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[140,1535],"filename":"addstreamer.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001426","name":"addstreamer","type":"ArrowFunctionExpression"},"vars":{"roles":"addstreamer~roles","channels":"addstreamer~channels","":null,"twitch_username":"addstreamer~twitch_username","discord_username":"addstreamer~discord_username"}},"description":"Execute on addstreamer command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"addstreamer","longname":"addstreamer","kind":"function","scope":"global","async":true,"$longname":"addstreamer","$kind":"method","$docmaLink":"?api#addstreamer"},{"comment":"/**\n * Execute on args command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[133,318],"filename":"args.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001627","name":"args","type":"ArrowFunctionExpression"}},"description":"Execute on args command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"args","longname":"args","kind":"function","scope":"global","async":true,"$longname":"args","$kind":"method","$docmaLink":"?api#args"},{"comment":"/**\n * Execute on avatar command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[93,704],"filename":"avatar.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001678","name":"avatar","type":"ArrowFunctionExpression"},"vars":{"avatarList":"avatar~avatarList","":null}},"description":"Execute on avatar command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"avatar","longname":"avatar","kind":"function","scope":"global","async":true,"$longname":"avatar","$kind":"method","$docmaLink":"?api#avatar"},{"comment":"/**\n * Execute on ban command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[90,232],"filename":"ban.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001774","name":"ban","type":"ArrowFunctionExpression"},"vars":{"user":"ban~user"}},"description":"Execute on ban command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"ban","longname":"ban","kind":"function","scope":"global","async":true,"$longname":"ban","$kind":"method","$docmaLink":"?api#ban"},{"comment":"/**\n * Execute on clear command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[92,824],"filename":"clear.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001829","name":"clear","type":"ArrowFunctionExpression"},"vars":{"deleteMessages":"clear~deleteMessages"}},"description":"Execute on clear command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"clear","longname":"clear","kind":"function","scope":"global","async":true,"$longname":"clear","$kind":"method","$docmaLink":"?api#clear"},{"comment":"/**\n * Execute on getdatabase command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[98,425],"filename":"getdatabase.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001938","name":"getdatabase","type":"ArrowFunctionExpression"},"vars":{"streamerDB":"getdatabase~streamerDB"}},"description":"Execute on getdatabase command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"getdatabase","longname":"getdatabase","kind":"function","scope":"global","async":true,"$longname":"getdatabase","$kind":"method","$docmaLink":"?api#getdatabase"},{"comment":"/**\n * Execute on getstreamers command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[141,1514],"filename":"getstreamers.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100001976","name":"getstreamers","type":"ArrowFunctionExpression"},"vars":{"roles":"getstreamers~roles","channels":"getstreamers~channels","":null,"twitch_username":"getstreamers~twitch_username","discord_username":"getstreamers~discord_username"}},"description":"Execute on getstreamers command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"getstreamers","longname":"getstreamers","kind":"function","scope":"global","async":true,"$longname":"getstreamers","$kind":"method","$docmaLink":"?api#getstreamers"},{"comment":"/**\n * Get User\n * @param {String} username - twitch username\n */","meta":{"range":[792,1201],"filename":"Twitch.js","lineno":31,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/utils","code":{"id":"astnode100006517","name":"getUser","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Get User","params":[{"type":{"names":["String"]},"description":"twitch username","name":"username"}],"name":"getUser","longname":"getUser","kind":"function","scope":"global","async":true,"$longname":"getUser","$kind":"method","$docmaLink":"?api#getUser"},{"comment":"/**\n * Execute on help command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[177,1793],"filename":"help.js","lineno":7,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002175","name":"help","type":"ArrowFunctionExpression"},"vars":{"data":"help~data","undefined":null,"":null,"name":"help~name","command":"help~command"}},"description":"Execute on help command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"help","longname":"help","kind":"function","scope":"global","async":true,"$longname":"help","$kind":"method","$docmaLink":"?api#help"},{"comment":"/**\n * Execute on invite command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[135,954],"filename":"invite.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002461","name":"invite","type":"ArrowFunctionExpression"},"vars":{"time":"invite~time","usage":"invite~usage","invite":"invite~invite"}},"description":"Execute on invite command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"invite","longname":"invite","kind":"function","scope":"global","async":true,"$longname":"invite","$kind":"method","$docmaLink":"?api#invite"},{"comment":"/**\n * Is Live\n * @param {Twitch.User} user - twitch user\n */","meta":{"range":[154,718],"filename":"Twitch.js","lineno":8,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/utils","code":{"id":"astnode100006445","name":"isLive","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Is Live","params":[{"type":{"names":["Twitch.User"]},"description":"twitch user","name":"user"}],"name":"isLive","longname":"isLive","kind":"function","scope":"global","async":true,"$longname":"isLive","$kind":"method","$docmaLink":"?api#isLive"},{"comment":"/**\n * Execute on kick command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[91,235],"filename":"kick.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002566","name":"kick","type":"ArrowFunctionExpression"},"vars":{"user":"kick~user"}},"description":"Execute on kick command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"kick","longname":"kick","kind":"function","scope":"global","async":true,"$longname":"kick","$kind":"method","$docmaLink":"?api#kick"},{"comment":"/**\n * Execute on message events\n * @param {Discord.Client} client - discord client\n */","meta":{"range":[218,5282],"filename":"message.js","lineno":9,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/actions","code":{"id":"astnode100000022","name":"message","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on message events","params":[{"type":{"names":["Discord.Client"]},"description":"discord client","name":"client"}],"name":"message","longname":"message","kind":"function","scope":"global","$longname":"message","$kind":"method","$docmaLink":"?api#message"},{"comment":"/**\n * Execute on mkcha command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,2630],"filename":"mkcha.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002621","name":"mkcha","type":"ArrowFunctionExpression"},"vars":{"channels":"mkcha~channels","categories":"mkcha~categories","":null,"channel_name":"mkcha~channel_name","channel_type":"mkcha~channel_type","parent_category":"mkcha~parent_category","shouldIBreak":"mkcha~shouldIBreak"}},"description":"Execute on mkcha command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"mkcha","longname":"mkcha","kind":"function","scope":"global","async":true,"$longname":"mkcha","$kind":"method","$docmaLink":"?api#mkcha"},{"comment":"/**\n * Execute on mute command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[91,570],"filename":"mute.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002886","name":"mute","type":"ArrowFunctionExpression"},"vars":{"roles":"mute~roles","":null,"user":"mute~user","muteRole":"mute~muteRole"}},"description":"Execute on mute command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"mute","longname":"mute","kind":"function","scope":"global","async":true,"$longname":"mute","$kind":"method","$docmaLink":"?api#mute"},{"comment":"/**\n * Execute on mvcha command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,2208],"filename":"mvcha.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100002994","name":"mvcha","type":"ArrowFunctionExpression"},"vars":{"channel_name":"mvcha~channel_name","new_channel_name":"mvcha~new_channel_name","new_category":"mvcha~new_category","channel_id":"mvcha~channel_id","":null,"givenCategory":"mvcha~givenCategory"}},"description":"Execute on mvcha command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"mvcha","longname":"mvcha","kind":"function","scope":"global","async":true,"$longname":"mvcha","$kind":"method","$docmaLink":"?api#mvcha"},{"comment":"/**\n * Execute on newrole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[136,4510],"filename":"newrole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003201","name":"newrole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"newrole~givenRole","permissions":"newrole~permissions","findRole":"newrole~findRole","":null,"options":"newrole~options"}},"description":"Execute on newrole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"newrole","longname":"newrole","kind":"function","scope":"global","async":true,"$longname":"newrole","$kind":"method","$docmaLink":"?api#newrole"},{"comment":"/**\n * Execute on pause command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[147,735],"filename":"pause.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003454","name":"pause","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"pause~serverQueue","serverQueue.playing":"pause~serverQueue.playing"}},"description":"Execute on pause command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"pause","longname":"pause","kind":"function","scope":"global","async":true,"$longname":"pause","$kind":"method","$docmaLink":"?api#pause"},{"comment":"/**\n * Execute on pin command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[132,563],"filename":"pin.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003575","name":"pin","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on pin command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"pin","longname":"pin","kind":"function","scope":"global","async":true,"$longname":"pin","$kind":"method","$docmaLink":"?api#pin"},{"comment":"/**\n * Execute on ping command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[91,153],"filename":"ping.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003671","name":"ping","type":"ArrowFunctionExpression"}},"description":"Execute on ping command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"ping","longname":"ping","kind":"function","scope":"global","async":true,"$longname":"ping","$kind":"method","$docmaLink":"?api#ping"},{"comment":"/**\n * Execute on play command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[1354,3616],"filename":"play.js","lineno":47,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003869","name":"play","type":"ArrowFunctionExpression"},"vars":{"url":"play~url","serverQueue":"play~serverQueue","voiceChannel":"play~voiceChannel","permissions":"play~permissions","songInfo":"play~songInfo","song":"play~song","queueConstract":"play~queueConstract","connection":"play~connection","queueConstract.connection":"play~queueConstract.connection"}},"description":"Execute on play command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"play","longname":"play","kind":"function","scope":"global","async":true,"$longname":"play","$kind":"method","$docmaLink":"?api#play"},{"comment":"/**\n * Play action\n * @param {Discord.Guild} guild - discord guild\n * @param {Queue} song - song queue\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[252,1218],"filename":"play.js","lineno":10,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100003715","name":"playAction","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"playAction~serverQueue","dispatcher":"playAction~dispatcher","":null}},"description":"Play action","params":[{"type":{"names":["Discord.Guild"]},"description":"discord guild","name":"guild"},{"type":{"names":["Queue"]},"description":"song queue","name":"song"},{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"playAction","longname":"playAction","kind":"function","scope":"global","$longname":"playAction","$kind":"method","$docmaLink":"?api#playAction"},{"comment":"/**\n * Execute on prune command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,636],"filename":"prune.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004176","name":"prune","type":"ArrowFunctionExpression"},"vars":{"amount":"prune~amount","":null}},"description":"Execute on prune command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"prune","longname":"prune","kind":"function","scope":"global","async":true,"$longname":"prune","$kind":"method","$docmaLink":"?api#prune"},{"comment":"/**\n * Execute on reaction events\n * @param {Discord.Client} client - discord client\n */","meta":{"range":[95,3004],"filename":"reactions.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/actions","code":{"id":"astnode100000590","name":"reactions","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on reaction events","params":[{"type":{"names":["Discord.Client"]},"description":"discord client","name":"client"}],"name":"reactions","longname":"reactions","kind":"function","scope":"global","$longname":"reactions","$kind":"method","$docmaLink":"?api#reactions"},{"comment":"/**\n * Execute on ready event\n * @param {Discord.Client} client - discord client\n */","meta":{"range":[176,2030],"filename":"ready.js","lineno":7,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/actions","code":{"id":"astnode100000887","name":"ready","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on ready event","params":[{"type":{"names":["Discord.Client"]},"description":"discord client","name":"client"}],"name":"ready","longname":"ready","kind":"function","scope":"global","$longname":"ready","$kind":"method","$docmaLink":"?api#ready"},{"comment":"/**\n * Execute on repeat command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[148,756],"filename":"repeat.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004270","name":"repeat","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"repeat~serverQueue","serverQueue.loop":"repeat~serverQueue.loop"}},"description":"Execute on repeat command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"repeat","longname":"repeat","kind":"function","scope":"global","async":true,"$longname":"repeat","$kind":"method","$docmaLink":"?api#repeat"},{"comment":"/**\n * Execute on resume command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[148,739],"filename":"resume.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004377","name":"resume","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"resume~serverQueue","serverQueue.playing":"resume~serverQueue.playing"}},"description":"Execute on resume command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"resume","longname":"resume","kind":"function","scope":"global","async":true,"$longname":"resume","$kind":"method","$docmaLink":"?api#resume"},{"comment":"/**\n * Execute on rmcha command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,699],"filename":"rmcha.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004497","name":"rmcha","type":"ArrowFunctionExpression"},"vars":{"channel_name":"rmcha~channel_name","channel_id":"rmcha~channel_id","":null,"fetchedChannel":"rmcha~fetchedChannel"}},"description":"Execute on rmcha command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"rmcha","longname":"rmcha","kind":"function","scope":"global","async":true,"$longname":"rmcha","$kind":"method","$docmaLink":"?api#rmcha"},{"comment":"/**\n * Execute on rmrole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[135,1179],"filename":"rmrole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004599","name":"rmrole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"rmrole~givenRole","member":"rmrole~member","findRole":"rmrole~findRole","":null}},"description":"Execute on rmrole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"rmrole","longname":"rmrole","kind":"function","scope":"global","async":true,"$longname":"rmrole","$kind":"method","$docmaLink":"?api#rmrole"},{"comment":"/**\n * Execute on server command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[93,243],"filename":"server.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004754","name":"server","type":"ArrowFunctionExpression"}},"description":"Execute on server command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"server","longname":"server","kind":"function","scope":"global","async":true,"$longname":"server","$kind":"method","$docmaLink":"?api#server"},{"comment":"/**\n * Execute on short command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[328,598],"filename":"short.js","lineno":10,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004825","name":"short","type":"ArrowFunctionExpression"},"vars":{"url":"short~url","response":"short~response"}},"description":"Execute on short command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"short","longname":"short","kind":"function","scope":"global","async":true,"$longname":"short","$kind":"method","$docmaLink":"?api#short"},{"comment":"/**\n * Execute on skip command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[146,687],"filename":"skip.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100004894","name":"skip","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"skip~serverQueue","serverQueue.loop":"skip~serverQueue.loop"}},"description":"Execute on skip command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"skip","longname":"skip","kind":"function","scope":"global","async":true,"$longname":"skip","$kind":"method","$docmaLink":"?api#skip"},{"comment":"/**\n * Execute on song command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[146,511],"filename":"song.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005021","name":"song","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"song~serverQueue"}},"description":"Execute on song command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"song","longname":"song","kind":"function","scope":"global","async":true,"$longname":"song","$kind":"method","$docmaLink":"?api#song"},{"comment":"/**\n * Execute on stop command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[146,721],"filename":"stop.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005113","name":"stop","type":"ArrowFunctionExpression"},"vars":{"serverQueue":"stop~serverQueue","serverQueue.songs":"stop~serverQueue.songs"}},"description":"Execute on stop command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"stop","longname":"stop","kind":"function","scope":"global","async":true,"$longname":"stop","$kind":"method","$docmaLink":"?api#stop"},{"comment":"/**\n * Execute on unban command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,716],"filename":"unban.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005246","name":"unban","type":"ArrowFunctionExpression"},"vars":{"username":"unban~username","":null}},"description":"Execute on unban command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"unban","longname":"unban","kind":"function","scope":"global","async":true,"$longname":"unban","$kind":"method","$docmaLink":"?api#unban"},{"comment":"/**\n * Execute on unmute command\n * @param {Discord.Message} msg - discord message\n */","meta":{"range":[93,581],"filename":"unmute.js","lineno":5,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005348","name":"unmute","type":"ArrowFunctionExpression"},"vars":{"roles":"unmute~roles","":null,"user":"unmute~user","muteRole":"unmute~muteRole"}},"description":"Execute on unmute command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"}],"name":"unmute","longname":"unmute","kind":"function","scope":"global","async":true,"$longname":"unmute","$kind":"method","$docmaLink":"?api#unmute"},{"comment":"/**\n * Execute on unpin command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[134,1976],"filename":"unpin.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005454","name":"unpin","type":"ArrowFunctionExpression"},"vars":{"":null}},"description":"Execute on unpin command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"unpin","longname":"unpin","kind":"function","scope":"global","async":true,"$longname":"unpin","$kind":"method","$docmaLink":"?api#unpin"},{"comment":"/**\n * Execute on virole command\n * @param {Discord.Message} msg - discord message\n * @param {Array} args - passed arguments\n */","meta":{"range":[135,9048],"filename":"virole.js","lineno":6,"columnno":6,"path":"/home/alotaima/Projects/side/SudoBot/app/commands","code":{"id":"astnode100005672","name":"virole","type":"ArrowFunctionExpression"},"vars":{"givenRole":"virole~givenRole","option":"virole~option","update":"virole~update","findRole":"virole~findRole","":null,"options":"virole~options","colorHelp":"virole~colorHelp","updateHelp":"virole~updateHelp","nameHelp":"virole~nameHelp"}},"description":"Execute on virole command","params":[{"type":{"names":["Discord.Message"]},"description":"discord message","name":"msg"},{"type":{"names":["Array"]},"description":"passed arguments","name":"args"}],"name":"virole","longname":"virole","kind":"function","scope":"global","async":true,"$longname":"virole","$kind":"method","$docmaLink":"?api#virole"}],"symbols":["addembedding","addproject","addrole","adds","addstreamer","args","avatar","ban","clear","getdatabase","getstreamers","getUser","help","invite","isLive","kick","message","mkcha","mute","mvcha","newrole","pause","pin","ping","play","playAction","prune","reactions","ready","repeat","resume","rmcha","rmrole","server","short","skip","song","stop","unban","unmute","unpin","virole"]}},"app":{"base":"/SudoBot/","title":"","meta":null,"entrance":"api","routing":{"method":"query","caseSensitive":true},"server":"static","favicon":""},"template":{"name":"docma-template-zebra","description":"Zebra - Default template for Docma. https://github.com/onury/docma","version":"2.3.1","supportedDocmaVersion":">=2.0.0","author":"Onur Yıldırım","license":"MIT","mainHTML":"index.html","options":{"title":{"label":"SudoBot","href":"/SudoBot/"},"logo":{"dark":"https://avatars.githubusercontent.com/u/72315676?s=200&v=4","light":"https://avatars.githubusercontent.com/u/72315676?s=200&v=4"},"sidebar":{"enabled":true,"outline":"tree","collapsed":false,"toolbar":true,"itemsFolded":false,"itemsOverflow":"crop","badges":true,"search":true,"animations":true},"symbols":{"autoLink":true,"params":"list","enums":"list","props":"list","meta":true},"contentView":{"bookmarks":"h1,h2,h3","faLibs":"solid,regular,brands","faVersion":"5.5.0"},"navbar":{"enabled":true,"dark":false,"animations":true,"menu":[{"label":"GitHub","iconClass":"fab fa-github","href":"https://github.com/sudogroup/SudoBot","target":"_blank"}],"fixed":true}}},"partials":{"api":"docma-api","content":"docma-content","notFound":"docma-404"},"elementID":"docma-main","contentElementID":"docma-content","defaultApiName":"_def_","logsEnabled":true}));
 
 /* global docma, DocmaWeb, page, sessionStorage */
 /* eslint no-nested-ternary:0, max-depth:0, no-var:0, prefer-template:0, prefer-arrow-callback:0, prefer-spread:0, object-shorthand:0 */
